Develop

+++ app/src/main/java/com/moritoui/vegegrowthapp/ui/home/HomeScreenViewModel.kt
@@ -2,25 +2,40 @@ package com.moritoui.vegegrowthapp.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.moritoui.vegegrowthapp.data.room.model.VegetableFolderEntity
 import com.moritoui.vegegrowthapp.model.FilterStatus
 import com.moritoui.vegegrowthapp.model.SelectMenu
 import com.moritoui.vegegrowthapp.model.VegeCategory
 import com.moritoui.vegegrowthapp.model.VegeItem
+import com.moritoui.vegegrowthapp.model.VegeItemDetail
 import com.moritoui.vegegrowthapp.model.VegeStatus
 import com.moritoui.vegegrowthapp.model.filterStatusMap
 import com.moritoui.vegegrowthapp.repository.datamigration.DataMigrationRepository
+import com.moritoui.vegegrowthapp.ui.home.model.AddDialogType
 import com.moritoui.vegegrowthapp.ui.home.model.HomeScreenUiState
+import com.moritoui.vegegrowthapp.ui.home.model.VegetablesState
 import com.moritoui.vegegrowthapp.usecases.AddVegeItemUseCase
 import com.moritoui.vegegrowthapp.usecases.ChangeVegeItemStatusUseCase
+import com.moritoui.vegegrowthapp.usecases.DeleteVegeFolderUseCase
 import com.moritoui.vegegrowthapp.usecases.DeleteVegeItemUseCase
-import com.moritoui.vegegrowthapp.usecases.GetVegeItemListUseCase
+import com.moritoui.vegegrowthapp.usecases.GetVegeItemDetailLastUseCase
+import com.moritoui.vegegrowthapp.usecases.GetVegeItemFromFolderIdUseCase
+import com.moritoui.vegegrowthapp.usecases.GetVegetableFolderUseCase
+import com.moritoui.vegegrowthapp.usecases.InsertVegetableFolderUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import java.util.UUID
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -29,24 +44,51 @@ import javax.inject.Inject
 class HomeScreenViewModel @Inject constructor(
     private val addVegeItemUseCase: AddVegeItemUseCase,
     private val deleteVegeItemUseCase: DeleteVegeItemUseCase,
-    private val getVegeItemListUseCase: GetVegeItemListUseCase,
+    private val deleteVegeFolderUseCase: DeleteVegeFolderUseCase,
+    private val getVegeItemFromFolderIdUseCase: GetVegeItemFromFolderIdUseCase,
     private val changeVegeItemStatusUseCase: ChangeVegeItemStatusUseCase,
-    private val dataMigrationRepository: DataMigrationRepository
+    private val getVegeItemDetailLastUseCase: GetVegeItemDetailLastUseCase,
+    private val dataMigrationRepository: DataMigrationRepository,
+    private val getVegetableFolderUseCase: GetVegetableFolderUseCase,
+    private val insertVegetableFolderUseCase: InsertVegetableFolderUseCase,
 ) : ViewModel() {
+    // ホーム画面では未分類のフォルダーIDnullのみを表示する
+    private val folderId = null
+
     private val _uiState = MutableStateFlow(HomeScreenUiState.initialState())
     val uiState: StateFlow<HomeScreenUiState> = _uiState.asStateFlow()
 
     private val _isDataMigrating = dataMigrationRepository.isDataMigrating
     val isDataMigrating: StateFlow<Boolean> = _isDataMigrating.asStateFlow()
 
+    private val _vegetables = MutableStateFlow<List<VegeItem>>(emptyList())
+    private val _vegetableDetails = MutableStateFlow<List<VegeItemDetail?>>(emptyList())
+    private val _vegetableFolders = MutableStateFlow<List<VegetableFolderEntity>>(emptyList())
+    val vegetablesState: StateFlow<VegetablesState> = combine(
+        _vegetables,
+        _vegetableDetails,
+        _vegetableFolders
+    ) { vegetables, vegetableDetails, vegetableFolders ->
+        VegetablesState(
+            vegetables = vegetables,
+            vegetableDetails = vegetableDetails,
+            vegetableFolders = vegetableFolders
+        )
+    }.stateIn(
+        viewModelScope,
+        SharingStarted.WhileSubscribed(5000),
+        VegetablesState.initial()
+    )
+
+    /**
+     * インサートエラーかどうか
+     */
+    private val _insertVegetableFolderEvent: MutableSharedFlow<Boolean> = MutableStateFlow(false)
+    val insertVegetableFolderEvent: SharedFlow<Boolean> = _insertVegetableFolderEvent.asSharedFlow()
+
     init {
         viewModelScope.launch {
             dataMigrationRepository.dataMigration()
-            _uiState.update {
-                it.copy(
-                    vegetables = getVegeItemListUseCase()
-                )
-            }
         }
         monitorUiState()
     }
@@ -54,15 +96,15 @@ class HomeScreenViewModel @Inject constructor(
     fun closeDialog() {
         _uiState.update {
             it.copy(
-                isOpenAddDialog = false
+                openAddDialogType = AddDialogType.NotOpenDialog
             )
         }
     }
 
-    fun openAddDialog() {
+    fun openAddDialog(addDialogType: AddDialogType) {
         _uiState.update {
             it.copy(
-                isOpenAddDialog = true,
+                openAddDialogType = addDialogType,
                 inputText = "",
                 selectCategory = VegeCategory.None
             )
@@ -70,17 +112,7 @@ class HomeScreenViewModel @Inject constructor(
     }
 
     fun selectStatus(vegeItem: VegeItem) {
-        viewModelScope.launch {
-            changeVegeItemStatusUseCase(vegeItem)
-            _uiState.value.vegetables.map { old ->
-                if (old.id == vegeItem.id) {
-                    vegeItem
-                } else {
-                    old
-                }
-            }
-            reloadVegetables()
-        }
+        changeVegeItem(vegeItem)
     }
 
     fun changeInputText(inputText: String) {
@@ -93,18 +125,48 @@ class HomeScreenViewModel @Inject constructor(
         }
     }
 
-    fun saveVegeItem() {
-        val vegeItem =
-            VegeItem(
-                name = _uiState.value.inputText,
-                category = _uiState.value.selectCategory,
-                uuid = UUID.randomUUID().toString(),
-                status = VegeStatus.Default
-            )
-        viewModelScope.launch {
-            addVegeItemUseCase(vegeItem)
-            reloadVegetables()
-            closeDialog()
+    /**
+     * 押されたダイアログの種類によって処理を変える
+     */
+    fun onAddDialogConfirm(addDialogType: AddDialogType) {
+        when (addDialogType) {
+            AddDialogType.AddVegeItem -> {
+                val vegeItem =
+                    VegeItem(
+                        name = _uiState.value.inputText,
+                        category = _uiState.value.selectCategory,
+                        uuid = UUID.randomUUID().toString(),
+                        status = VegeStatus.Default,
+                        folderId = null
+                    )
+                viewModelScope.launch {
+                    addVegeItemUseCase(vegeItem)
+                    reloadVegetables()
+                    closeDialog()
+                }
+            }
+            AddDialogType.AddFolder -> {
+                val vegeFolder = VegetableFolderEntity(
+                    id = 0,
+                    folderNumber = _vegetableFolders.value.size,
+                    folderName = _uiState.value.inputText,
+                    vegetableCategory = _uiState.value.selectCategory
+                )
+                viewModelScope.launch {
+                    insertVegetableFolderUseCase(vegeFolder)
+                        .onSuccess {
+                            reloadVegetables()
+                            closeDialog()
+                        }.onFailure {
+                            _insertVegetableFolderEvent.emit(true)
+                            delay(2000)
+                            _insertVegetableFolderEvent.emit(false)
+                        }
+                }
+            }
+            else -> {
+                return
+            }
         }
     }
 
@@ -134,12 +196,34 @@ class HomeScreenViewModel @Inject constructor(
         }
     }
 
+    fun changeFolderMoveMode() {
+        _uiState.update {
+            it.copy(
+                selectMenu =
+                if (_uiState.value.selectMenu == SelectMenu.Edit) {
+                    SelectMenu.None
+                } else {
+                    SelectMenu.MoveFolder
+                }
+            )
+        }
+    }
+
     fun deleteItem() {
-        val deleteItem = _uiState.value.targetDeleteItem ?: return
         viewModelScope.launch {
-            deleteVegeItemUseCase(deleteItem)
+            val deleteItem = _uiState.value.selectedItem
+            val deleteFolder = _uiState.value.selectedFolder
+            if (deleteItem != null) {
+                deleteVegeItemUseCase(deleteItem)
+            }
+            if (deleteFolder != null) {
+                deleteVegeFolderUseCase(deleteFolder)
+            }
             _uiState.update {
-                it.copy(targetDeleteItem = null)
+                it.copy(
+                    selectedItem = null,
+                    selectedFolder = null
+                )
             }
         }
     }
@@ -169,6 +253,32 @@ class HomeScreenViewModel @Inject constructor(
         }
     }
 
+    // フォルダ選択モーダルを開く
+    fun openFolderMoveBottomSheetState(vegeItem: VegeItem) {
+        _uiState.update {
+            it.copy(
+                isOpenFolderMoveBottomSheet = true,
+                selectedItem = vegeItem
+            )
+        }
+    }
+
+    // フォルダ選択モーダルを閉じる
+    fun closeFolderMoveBottomSheetState() {
+        _uiState.update {
+            it.copy(
+                isOpenFolderMoveBottomSheet = false
+            )
+        }
+    }
+
+    /**
+     * フォルダ移動をする
+     */
+    fun vegeItemMoveFolder(vegeItem: VegeItem) {
+        changeVegeItem(vegeItem)
+    }
+
     /**
      * 削除ダイアログを閉じる
      */
@@ -183,11 +293,23 @@ class HomeScreenViewModel @Inject constructor(
     /**
      * 削除ダイアログを開いて，削除するものをセットする
      */
-    fun openDeleteDialog(vegeItem: VegeItem) {
+    fun openDeleteVegeItemDialog(vegeItem: VegeItem) {
         _uiState.update {
             it.copy(
                 isOpenDeleteDialog = true,
-                targetDeleteItem = vegeItem
+                selectedItem = vegeItem
+            )
+        }
+    }
+
+    /**
+     * 削除ダイアログを開いて，削除するものをセットする
+     */
+    fun openDeleteFolderDialog(folder: VegetableFolderEntity) {
+        _uiState.update {
+            it.copy(
+                isOpenDeleteDialog = true,
+                selectedFolder = folder
             )
         }
     }
@@ -197,32 +319,69 @@ class HomeScreenViewModel @Inject constructor(
         else -> true
     }
 
-    private fun reloadVegetables() {
+    /**
+     * 登録されている野菜のリストを更新する
+     */
+    fun reloadVegetables() {
         viewModelScope.launch {
             val filterStatus = _uiState.value.filterStatus
-            val filteredVegetables =
-                getVegeItemListUseCase().filter { item ->
-                    if (filterStatus == FilterStatus.All) {
-                        true
-                    } else {
-                        item.status == filterStatusMap[filterStatus] ||
-                                item.category == filterStatusMap[filterStatus]
-                    }
+            val filteredVegetables = getVegeItemFromFolderIdUseCase(folderId).filter { item ->
+                if (filterStatus == FilterStatus.All) {
+                    true
+                } else {
+                    item.status == filterStatusMap[filterStatus] ||
+                        item.category == filterStatusMap[filterStatus]
                 }
-            _uiState.update {
-                it.copy(
-                    vegetables = filteredVegetables
-                )
+            }
+
+            _vegetables.update {
+                filteredVegetables
+            }
+
+            _vegetableDetails.update {
+                filteredVegetables.map { reloadVegetableDetailLast(it) }
+            }
+
+            _vegetableFolders.update {
+                getVegetableFolderUseCase()
             }
         }
     }
+
+    /**
+     * 指定された野菜の最新登録情報を取得する
+     */
+    private suspend fun reloadVegetableDetailLast(vegeItem: VegeItem): VegeItemDetail? = getVegeItemDetailLastUseCase(vegeItem.id)
 
     /**
      * uiStateの変更を監視する
      */
     private fun monitorUiState() {
         _uiState.onEach {
+            _uiState.update {
+                it.copy(isLoading = true)
+            }
             reloadVegetables()
+            _uiState.update {
+                it.copy(isLoading = false)
+            }
         }.launchIn(viewModelScope)
+    }
+
+    /**
+     * 登録されている野菜の情報を更新する
+     */
+    private fun changeVegeItem(vegeItem: VegeItem) {
+        viewModelScope.launch {
+            changeVegeItemStatusUseCase(vegeItem)
+            _vegetables.value.map { old ->
+                if (old.id == vegeItem.id) {
+                    vegeItem
+                } else {
+                    old
+                }
+            }
+            reloadVegetables()
+        }
     }
 }
